Classes and interfaces moved to 'json' package because it's not need a separate package
