+++ src/main/java/redis/clients/jedis/json/JsonObjectMapper.java
@@ -1,4 +1,6 @@
-package redis.clients.jedis.json.parser;
+package redis.clients.jedis.json;
+
+import redis.clients.jedis.json.DefaultGsonObjectMapper;
 
 /**
  * Represents the ability of serialize an object to JSON format string and deserialize it to the
