--- src/main/java/redis/clients/jedis/json/parser/package-info.java
@@ -1,4 +1,0 @@
-/**
- * This package contains the classes and interfaces related to JSON processing module.
- */
-package redis.clients.jedis.json.parser;
